requirements.txt
Django==4.0.6
MarkupPy==1.14
Pillow==9.2.0
PyYAML==6.0
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
django-import-export==2.8.0
django-material-admin==1.8.6
et-xmlfile==1.1.0
odfpy==1.4.1
openpyxl==3.0.10
pip==22.0.4
pytz==2022.2.1
setuptools==58.1.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.2
xlrd==2.0.1
xlwt==1.3.0